+++ packages/react/src/themes/default-theme/index.ts
@@ -1,0 +1,2 @@
+export * from './default-theme';
+export * from './default-component-tokens';
