+++ packages/react/src/themes/interface/index.ts
@@ -3,4 +3,3 @@ export { TextAttributes, TextAttributeKeys } from './text-attributes';
 export { RefTokens, RefTokenKeys } from './ref-tokens';
 export { AliasTokens, AliasTokenKeys } from './alias-tokens';
 export { ComponentTokens, ComponentTokensKeys } from './component-tokens';
-export { defaultComponentTokens } from './tokens-generator';
